+++ compiler-core/src/metadata.rs
@@ -445,6 +445,15 @@ impl RemapIds {
                 type_: self.type_(type_),
                 extra_information,
             },
+            Constant::Todo {
+                location,
+                type_,
+                message,
+            } => Constant::Todo {
+                location,
+                type_: self.type_(type_),
+                message: message.map(|message| Box::new(self.constant(*message))),
+            },
         }
     }